requirements.txt
aiohttp==3.3.0
astroid==2.0.3
async-timeout==3.0.0
chardet==3.0.4
colorama==0.3.9
discord.py==1.2.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.3.1
pylint==2.1.1
six==1.11.0
typed-ast==1.1.0
websockets==6.0
wrapt==1.10.11
yarl==1.1.1
praw==6.3.1
peewee==3.9.6
